better build script

+++ NASA-TM/build.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 #build pdf (with bibtex)
 "/usr/texbin/pdflatex" -synctex=1 -shell-escape -interaction=nonstopmode heading.tex
-bibtex heading.aux
-makeindex heading.nlo  -s nomencl.ist -o heading.nls
-"/usr/texbin/pdflatex" -synctex=1 -shell-escape -interaction=nonstopmode heading.tex
-"/usr/texbin/pdflatex" -synctex=1 -shell-escape -interaction=nonstopmode heading.tex
 
-mv heading.pdf hyperloop.pdf
+if [ "$1" == "-f" ]; then #add -f flag to do full build
+	bibtex heading.aux
+	makeindex heading.nlo  -s nomencl.ist -o heading.nls
+	"/usr/texbin/pdflatex" -synctex=1 -shell-escape -interaction=nonstopmode heading.tex
+	"/usr/texbin/pdflatex" -synctex=1 -shell-escape -interaction=nonstopmode heading.tex
+
+	mv heading.pdf hyperloop.pdf
+fi
+
 
 #. clean.sh
